+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/PartyService.cs
@@ -20,7 +20,6 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using Newtonsoft.Json.Linq;
 using Stormancer.Core;
 using Stormancer.Diagnostics;
 using Stormancer.Plugins;
@@ -166,6 +165,7 @@ namespace Stormancer.Server.Plugins.Party
 
         internal async Task OnConnecting(IScenePeerClient peer)
         {
+            var handlers = _handlers;
             await _partyState.TaskQueue.PushWork(async () =>
             {
                 //Todo jojo later
@@ -190,7 +190,7 @@ namespace Stormancer.Server.Plugins.Party
                 }
 
                 var ctx = new JoiningPartyContext(this, session, _partyState.PendingAcceptedPeers.Count + _partyState.PartyMembers.Count);
-                await _handlers().RunEventHandler(h => h.OnJoining(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnJoining", ex));
+                await handlers().RunEventHandler(h => h.OnJoining(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnJoining", ex));
                 if (!ctx.Accept)
                 {
                     Log(LogLevel.Trace, "OnConnecting", "Join denied by event handler", peer.SessionId, session.User?.Id);
@@ -339,17 +339,23 @@ namespace Stormancer.Server.Plugins.Party
             }
         }
 
-        public async Task UpdateSettings(PartySettingsDto partySettingsDto)
+        public async Task UpdateSettings(PartySettingsDto partySettingsDto, CancellationToken ct)
         {
+            var handlers = _handlers;
             await _partyState.TaskQueue.PushWork(async () =>
             {
+                if (ct.IsCancellationRequested)
+                {
+                    return;
+                }
+
                 if (partySettingsDto.GameFinderName == "")
                 {
                     throw new ClientException(GameFinderNameError);
                 }
                 var originalDto = partySettingsDto.Clone();
                 var ctx = new PartySettingsUpdateCtx(this, partySettingsDto);
-                await _handlers().RunEventHandler(h => h.OnUpdatingSettings(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnUpdatingSettings", ex));
+                await handlers().RunEventHandler(h => h.OnUpdatingSettings(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnUpdatingSettings", ex));
 
                 if (!ctx.ApplyChanges)
                 {
@@ -358,7 +364,7 @@ namespace Stormancer.Server.Plugins.Party
                 }
 
                 Log(LogLevel.Trace, "UpdateSettings", "Settings update accepted", partySettingsDto);
-                await _handlers().RunEventHandler(h => h.OnUpdateSettings(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnUpdateSettings", ex));
+                await handlers().RunEventHandler(h => h.OnUpdateSettings(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnUpdateSettings", ex));
 
                 // If the event handlers have modified the settings, we need to notify the leader to invalidate their local copy.
                 // Make an additional bump to the version number to achieve this.
@@ -382,7 +388,7 @@ namespace Stormancer.Server.Plugins.Party
 
                 Dictionary<PartyMember, PartySettingsUpdateDto> updates = _partyState.PartyMembers.Values.ToDictionary(m => m, _ => new PartySettingsUpdateDto(_partyState));
 
-                await _handlers().RunEventHandler(h => h.OnSendingSettingsUpdateToMembers(new PartySettingsMemberUpdateCtx(this, updates)),
+                await handlers().RunEventHandler(h => h.OnSendingSettingsUpdateToMembers(new PartySettingsMemberUpdateCtx(this, updates)),
                 ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnSendingSettingsToMember", ex));
 
                 await BroadcastStateUpdateRpc(
@@ -398,7 +404,7 @@ namespace Stormancer.Server.Plugins.Party
         /// <param name="partyUserStatus"></param>
         /// <param name="userId"></param>
         /// <returns></returns>
-        public async Task UpdateGameFinderPlayerStatus(string userId, PartyMemberStatusUpdateRequest partyUserStatus)
+        public async Task UpdateGameFinderPlayerStatus(string userId, PartyMemberStatusUpdateRequest partyUserStatus, CancellationToken ct)
         {
             if (string.IsNullOrEmpty(userId))
             {
@@ -410,8 +416,13 @@ namespace Stormancer.Server.Plugins.Party
                 throw new ArgumentNullException(nameof(partyUserStatus));
             }
 
+            var handlers = _handlers;
             await _partyState.TaskQueue.PushWork(async () =>
             {
+                if (ct.IsCancellationRequested)
+                {
+                    return;
+                }
 
                 if (!TryGetMemberByUserId(userId, out var user))
                 {
@@ -436,7 +447,7 @@ namespace Stormancer.Server.Plugins.Party
                 await BroadcastStateUpdateRpc(BatchStatusUpdate.Route, update);
 
                 var eventHandlerCtx = new PlayerReadyStateContext(this, user);
-                await _handlers().RunEventHandler(h => h.OnPlayerReadyStateChanged(eventHandlerCtx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnPlayerReadyStateChanged", ex));
+                await handlers().RunEventHandler(h => h.OnPlayerReadyStateChanged(eventHandlerCtx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnPlayerReadyStateChanged", ex));
 
                 bool shouldLaunchGameFinderRequest = false;
                 switch (eventHandlerCtx.GameFinderPolicy)
@@ -699,17 +710,23 @@ namespace Stormancer.Server.Plugins.Party
             _scene.Broadcast(route, data);
         }
 
-        public async Task SendPartyState(string recipientUserId)
+        public async Task SendPartyState(string recipientUserId, CancellationToken ct)
         {
+            var handlers = _handlers;
             await _partyState.TaskQueue.PushWork(async () =>
             {
+                if (ct.IsCancellationRequested)
+                {
+                    return;
+                }
+
                 PartyMember member;
                 if (!TryGetMemberByUserId(recipientUserId, out member))
                 {
                     throw new ClientException(NoSuchMemberError(recipientUserId));
                 }
 
-                var state = await MakePartyStateDto(member);
+                var state = await MakePartyStateDto(member, handlers);
 
                 using (var cts = new CancellationTokenSource())
                 {
@@ -728,11 +745,17 @@ namespace Stormancer.Server.Plugins.Party
 
         public async Task SendPartyStateAsRequestAnswer(RequestContext<IScenePeerClient> ctx)
         {
+            var handlers = _handlers;
             await _partyState.TaskQueue.PushWork(async () =>
             {
+                if (ctx.CancellationToken.IsCancellationRequested)
+                {
+                    throw new TaskCanceledException();
+                }
+
                 if (_partyState.PartyMembers.TryGetValue(ctx.RemotePeer.SessionId, out var member))
                 {
-                    var dto = await MakePartyStateDto(member);
+                    var dto = await MakePartyStateDto(member, handlers);
 
                     return ctx.SendValue(dto);
                 }
@@ -743,7 +766,7 @@ namespace Stormancer.Server.Plugins.Party
             });
         }
 
-        private async Task<PartyStateDto> MakePartyStateDto(PartyMember recipient)
+        private async Task<PartyStateDto> MakePartyStateDto(PartyMember recipient, Func<IEnumerable<IPartyEventHandler>> handlers)
         {
             var dto = new PartyStateDto
             {
@@ -754,7 +777,7 @@ namespace Stormancer.Server.Plugins.Party
                 Version = _partyState.VersionNumber
             };
 
-            await _handlers().RunEventHandler(
+            await handlers().RunEventHandler(
                 h => h.OnSendingSettingsUpdateToMembers(new PartySettingsMemberUpdateCtx(this, new Dictionary<PartyMember, PartySettingsUpdateDto> { { recipient, dto.Settings } })),
                 ex => Log(LogLevel.Error, "MakePartyStateDto", "An exception was thrown by a OnSendingSettingsUpdateToMembers handler"));